MDCACHE - Don't update attributes when they're invalid
If an entry has had it's attributes invalidated, and an update() UP call
arrives, don't update the attributes and mark them valid, since this
will likely just be a subset of the full attributes.  Instead, just bail
from the update() call, and allow the attributes to be refreshed on
demand.

Change-Id: I3c5cfd0d1a01c839769db0b5a8bb1884ff18450e
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>