Implement an LRU list for loose dirents
A new config parameter Loose_Mult limits the number of loose dirents.
The maximum is expressed as a multiple of the chunk size (allowing
most instances to leave this option at default).

As loose dirents are added, they are added to an LRU list in the MRU
position. If the list already contains the maximum number, the LRU
item will be discarded. Any lookup that finds a loose dirent in cache
will move that entry to the MRU position in the list.

Change-Id: Iffa0662ce610b2aa978deb1922d37fd11fadcd9a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>