WIP - Implement an LRU list for detached dirents
WIP NOTE - This needs some work, the LRU list needs a write lock
to be able to bump entries, or a different mechanism used.

A new config parameter Detached_Mult limits the number of detached
dirents. The maximum is expressed as a multiple of the chunk size
(allowing most instances to leave this option at default).

As detached dirents are added, they are added to an LRU list in the
MRU position. If the list already contains the maximum number, the LRU
item will be discarded. Any lookup that finds a detached dirent in
cache will move that entry to the MRU position in the list.

Change-Id: Iffa0662ce610b2aa978deb1922d37fd11fadcd9a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>