Use care_t care instead of bool nsm_state_applies in get_nlm_state
We need to create a new state_t for non-monitored requests
(NLM4_NM_LOCK and NLM4_SHARE), and since care_t already communicates
almost everything, we just pass it. We add a new value CARE_OWNER for
NLM4_TEST which does need to create an owner, but doesn't need to
create a state_t if there isn't one present.

Change-Id: I525dbb368215955015941851a8dc14eb4524b47a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>