Add new Reaper_Work_Per_Lane option
The Reaper_Work option was only used to set work-per-lane within the LRU
thread.  However, if the number of lanes is increased (as one company
does), this can result in only looking at one handle per lane, resulting
in many (most?) runs not demoting any handles.

Instead, add a new option Reaper_Work_Per_Lane that explicitly sets the
number of handles to check per-lane.  This should work, regardless of
the number of lanes.

Deprecate Reaper_Work, and change it's default to 0 (unused).  If it's
set, it will still be used.

Change-Id: I146edf953b2d5d1ab0d33a9a5c8bd4732fd9df24
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>