Fix EID6h and EID7 pynfs 4.1 test cases
Fix Case 6 - Update. We don't actually manage any of the fields that
can be updated, but we can also just quietly accept the update.

We need to reject eia_flags value EXCHGID4_FLAG_CONFIRMED_R on
an EXCHANGE_ID request, it's a server only flag.

Also fix a refcount leak.

Change-Id: Ie62a362424ecdb5594134a25838f8b9dc6e26ed1
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>