nfs: fix error handling in layoutrecall code
We need to free any FH data on error here, but at the point where we
call gsh_free on it in layoutrecall, that pointer will likely just
have junk in it.

Conversely, in the event that the call submission fails, we currently
just leak the FH data. Fix the code to free it appropriately in that
case.

Change-Id: I3c8910bc9f2f1a024a3b4b83f691f2e0eaf56118
Signed-off-by: Jeff Layton <jlayton@redhat.com>