Update patch set 1
Patch Set 1:

I'd suggest merging this patch ahead of the others since it seems like a straightforward bugfix. There is one hunk in here that might be good to break out into into a separate patch, however, mostly just so we can get a clear description of the problem it fixes.

Patch-set: 1