Rework delegations to use new lease_op2 interface
This patch is heavily based on Soumya Koduri's original work to add a
new lease_op2 operation. The original proposal had quite a bit more
fields to handle reclaims, and different fields to handle setting
and removing and whether it was a read only delegation.

Ganesha never reclaims delegations, so I don't think we really need
that setting at this point. Also, the r/w setting is ignored when
we're removing a delegation. That leaves us with only 3 different
states for lease_op2 -- no delegation, read delegation, or read/write
delegation.

This proposed lease_op2 interface just takes an owner pointer and a
fsal_deleg_t enum, that is one of FSAL_DELEG_NONE, FSAL_DELEG_RD,
FSAL_DELEG_WR. I think this covers all of the use cases and is simpler
for FSAL's to get right.

Change-Id: I9ef314125c5eb86fd21c449787349b78f32da1df
Signed-off-by: Soumya Koduri <skoduri@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>