Fix looping in release_lock_owner
It is possible to have multiple states with the same owner and export.
When this happens, only the last state will exist in the ht_state_obj
hash table.

Calls to nfs4_State_Del() return false for all but the last state, and
the state is not removed from the owner's so_state_list. This causes
release_lock_owner to go into an infinite loop!

Fixed nfs4_State_Del() to return true as long as it exists in
ht_state_id hash table and we successfully remove from it.

Change-Id: If27e678d38a3df02cbd9e71873a2921e7eb46368
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>