Use 'v6disabled' flag to know if IPv6 is disabled
Currently ganesha startup on a system with IPv6 disabled fails
with message "Register_program :DISP :FATAL :Cannot register
NFS V3 on UDPv6". This is because there is no check to
know if IPv6 is enabled or not in Register_Program to decide if
a protocol needs to be registered for IPv6 or not.
Fixed this by checking 'v6disabled' flag to know if IPv6 is
disabled. If yes, in Register_Program we do UDP_REGISTER/TCP_REGISTER
for netconfig_udp4/netconfig_tcp4. Else we check if
netconfig is enabled for udp6/tcp6. If no in Register_Program
we do UDP_REGISTER/TCP_REGISTER for netconfig_udp4/netconfig_tcp4,
else in Register_Program we do UDP_REGISTER/TCP_REGISTER
for netconfig_udp6/netconfig_tcp6.

Change-Id: I9e9aa6b89cf224bcf9623bf3200d43854724ab61
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>