FSAL_GLUSTER : use fd present in open state_t when ever available
Currently FSAL_GLUSTER opens two different fd's for open state_t and lock_t.
There will be sceneraios in which open was successful but before sending
lock request the file might be removed and as per posix behaviour the lock
request already opened fd should not fail with ENOENT, and application may
expect so. Solution is to use same fd's for both open and lock call

Change-Id: I3b8ce4a47541ec9fbada2166a54707e6ccc4ee69
Signed-off-by: Jiffin Tony Thottan <jthottan@redhat.com>