FSAL_GLUSTER: Use the APIs to register for upcall events' callback
There are new APIs being provided by libgfapi to register
and unregister for various upcall event along with callbacks
functions to be invoked  - https://review.gluster.org/18349

So instead of polling continuously in a loop hogging CPU,
use these new APIs to register callback functions for cache
invalidation. These callback functions shall be invoked
only in the event of any upcall received and that too in the
backend fs thread context (which seem to be safe as per the current
upcall design).

Change-Id: I46b103d3b8555116efe5852eab99de30a5fd0e43
Signed-off-by: Soumya Koduri <skoduri@redhat.com>