MDCACHE - Release unused new entries
When a new entry is allocated, it can be left unused (not put into the
hash table or LRU), and released after that.  However,
mdcache_kill_entry() would only try to release if it was in the hash
table.

Instead of calling mdcache_kill_entry(), just put the sentinal ref.

Change-Id: I93fb9fb2e264ca02fbe965cb4cd97b81bde2ce68
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>