drc: cleanup duplicate checking of refcnt in drc_free_expired
As described at "The life of tcp drc", when a drc's refcount goes to zero,
it is put onto recycle queue.
So only drc's with non-zero refcount exists in the tree alone.
Drc's with zero refcount should be in the tree as well as the queue.

All checking of "refcnt == 0" are duplidate in drc_free_expired().

Change-Id: Ie31cae5fe929a2d62849fc35a30a6b5b866b5906
Signed-off-by: Kinglong Mee <kinglongmee@gmail.com>