scripts: Multilock test client for Gluster
This contains the test client for Glusterfs which makes
use of gfapi library calls. Most of the tests have passed
but I observed few failures wrt to lock split specific to
LOCK,UNLOCK & LOCK cases. But the same failures are observed
on glusterfs fuse native client as well. So it needs to
be debugged on glusterfs server side if there is any issue
wrt to locks merge in case of add,subtract and add sequence.

Change-Id: I86b0ae15b4a3f87e8447fba804fec20668e40c28
Signed-off-by: Soumya Koduri <skoduri@redhat.com>