[GPFS] workaround for lookup of parent entry
Sometimes we receive invalid handle from lookup of "..", handle
it by sending DELAY error and hope it goes away in a retry!

Also, We shouldn't be sending a lookup of ".." in an export root or root
of the GPFS file system. If we do ever send, we get an ENOENT from GPFS,
so handle it gracefully.

Change-Id: I021b4a88542395a746b05242dbbbc92fbedd32b0
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>