Fix attr and content lock deadlock with readdir and other operations
attr_lock and content_lock of an object are well ordered.  Attr_lock
needs to be acquired first and then the content_lock. Readdir operation
acquires the content_lock on the directory and then attr_lock of a
child is acquired. The exception is for ".." direntry which actually
ends up acquiring attr_lock on its parent which leads to ABBA deadlock.

The fix drops the content_lock on the directory while searching for ".."
direntry. This preserves the order that parent object locks are acquired
before acquiring child locks, there by avoiding the deadlock.

Change-Id: I64a6ac7e8141179bcda2ca15dfab3912f808a74a
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
[for minor corrections and git message]