Replace "thread name" with request number for workers
Currently, the actual work is not done by the traditional worker
threads, but is instead done by ntirpc threads. These treads aren't
named for logging purposes.

This patch adds a capability to register a callback function with the
logging facility such that if the "thread name" is needed, and isn't
provided, a callback will be made to set the function name.

The function that sets the callback also clears the current function
name.

This allows the logging facility, if it will actually use the function
name to request it be generated.

The nfs_rpc_process_request function calls the new SetNameFunctionCB
at the start of each reqest processing. It also increments a global
request number, saving off it's request number.

When needed by the logging facility, the callback is used to generate
a function name of RN_xxxxx.

Change-Id: I980302049b140ac25ad9e9f61d9e85d23f9a0f71
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>