SAL: allow grace period to be lifted early if all clients have sent RECLAIM_COMPLETE
Keep track of the number of clients on the recovery list, and the number
on the list that have sent a RECLAIM_COMPLETE. v4.0 clients will never
send a RECLAIM_COMPLETE, so when those values are equal we can lift the
grace period immediately.

For now, this ignores v3 altogether. We'll probably want this to let
the grace period run to completion if any v3 SM_NOTIFYs were sent.

Maybe we should just base that on whether any exports have v3 enabled?

Also, this ignores any sort of clustering. We'll eventually need to
define a set of operations around the grace period handling so that
we can handle clustered configurations correctly.

Change-Id: Ie720da438220f7115d4458b4b995753d480d6491
Signed-off-by: Jeff Layton <jlayton@redhat.com>