Avoid multiple states for the same owner and object
If a client sends multiple requests with the same owner and create(open)
operation with the same name, we end up creating multiple states. This
makes ganesha unable to remove states from the owner list except the
last state as nfs4_State_Del() fails for those states (don't exits in
the ht_state_obj hashtable). This leads to an infinite loop in
release_openstate.

Replaced SET_OVERWRITE with NO_OVERWRITE which will detect this and
avoids the situation.

Change-Id: Ia0774280ca534bea3efa222e4c544625c8785f9c
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>