recovery_fs: rework db handling scheme for better fault-tolerance
The current recovery scheme is fiddly and doesn't tolerate faults well.
If you restart ganesha while it's in its grace period, you can lose
recovery entries. It's also susceptible to crashes that occur when
trying to load the entries from disk. It opens two different databases
and copies entries between them. That can fail midstream, and we end up
with a mess.

The key to fixing this is recognizing that that no client can set new
state until the grace period is lifted. There is no benefit to ensuring
the durability of new client records that did not exist in the previous
database until then. The pre-crash database is always authoritative
until that point.

Rework the code around this principle:

When we start ganesha, create a new empty temporary directory with a
unique name to record incoming clients. We'll always create new client
records in this directory, both for brand new clients and those
performing reclaim.

Reclaim of previous state however must be gated on the previous recovery
db, so we read those records into memory when starting . We do _not_
however preemptively create new entries for these records in the new
recovery dir. That will be done naturally as the clients connect to
reclaim their state.

Once we're ready to lift the grace period, we must atomically switch to
the new database from the old, so that reclaim after a subsequent crash
will use the new db. This must be done before handing out new state.

This is tricky as the recovery_fs driver uses a directory here,
and we can't rename one on top of another. What we do instead is
create a symlink that points to the new recovery database and then
rename that symlink on top of the old one (if any).

This scheme naturally handles the case where the server crashes within its
grace period. That can happen an arbitrary number of times and we can
still allow clients to reclaim until we reach the end of a grace period.

For now, we do not handle the takeover case. That is doable, but will
take a bit more work. It also doesn't handle the transition from the
old directory handling scheme.

Change-Id: I886b310aa74cff356dd30ff44e8a63941ed579bd
Signed-off-by: Jeff Layton <jlayton@redhat.com>