SAL: add new recovery_fs_ng backend that is simpler and more fault-tolerant
The current recovery scheme is fiddly and doesn't tolerate faults well.
If you restart ganesha while it's in its grace period, you can lose
recovery entries.

The key to fixing this is recognizing that that no client can set new
state until the grace period is lifted. There is no benefit to ensuring
the durability of new client records that did not exist in the previous
database until then. The pre-crash database is always authoritative
until that point.

Add a new recovery backend driver based on this principle. It shares
the basic storage format and much of the code from the legacy fs
backend, but manages the databases a little differently.

When we start ganesha, create a new empty temporary directory with a
unique name to record incoming clients. We'll always create new client
records in this directory, both for brand new clients and those
performing reclaim.

Reclaim of previous state however must be gated on the previous recovery
db, so we read those records into memory when starting. Do not, however
preemptively create new entries for these records in the new
recovery dir like the old code did. That will be done naturally as the
clients reconnect to reclaim their state.

Once we're ready to lift the grace period, we must atomically switch to
the new database from the old, so that reclaim after a subsequent crash
will use the new db. This must be done before handing out new state.

The only way to do that atomically is to rename and we can't rename
a directory on top of another. Instead, create a symlink that points to
the new recovery database and then rename that symlink on top of the old
one (if any). When we open the db for recovery, we just use the symlink.

This scheme naturally handles the case where the server crashes within its
grace period. That can happen an arbitrary number of times and we can
still allow clients to reclaim until we reach the end of a grace period.

For now, we do not handle the takeover case. That is doable, but will
take a bit more work.

Change-Id: I886b310aa74cff356dd30ff44e8a63941ed579bd
Signed-off-by: Jeff Layton <jlayton@redhat.com>