Add a global parameter to avoid dropping caps
Dropping caps is a useful thing for Ganesha to do, but when it's envoked
as a library, it's probably best to leave cap management to the main
program.

Add a global parameter to bypass dropping caps.  The default for
standalone Ganesha is to drop the caps; the default for library Ganesha
is to skip dropping caps.

Change-Id: I9a9100ca0f2de94651599630e1a4a32c0675dd24
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>