recovery_rados_ng: add new driver for recovery
Add a new recovery backend driver that is "safe by default".

We always load the client database from the on-disk store. As new
clients come in and leave during the grace period, we add and remove
them from a single shared "grace_op" transaction.

Once the grace period is lifted, we commit that transaction and
further updates to the kv store are done synchronously.

This method makes it OK for ganesha to crash during the grace period,
as changes to the backing store are done in memory until the grace period
is lifted, at which point the changes are committed atomically.

Change-Id: Ibc36f626df0b9e423b225726e2ada364c2eaed00
Signed-off-by: Jeff Layton <jlayton@redhat.com>