NFS: rework grace period handling
The current method leaves open the potential for races. We also need to
coordinate the lifting of the grace period with the cleanup/cutover of
the recovery db.

Protect changes to current_grace with the grace_mutex, but allow
atomic stores when changing it, and atomic fetches against it
for grace checks without the mutex. That allows us to keep the
grace_mutex out of the codepath during normal operation.

When we go to lift the grace period clean out the recovery db, and
set current_grace to 0. current_grace of 0 signifies that the grace
period is now over.

Change-Id: Ia396978971d4a07ff42fd9dbd5da65c1af0450ba
Signed-off-by: Jeff Layton <jlayton@redhat.com>