NFS: rework grace period handling
The current method leaves open the potential for races. We also need to
coordinate the lifting of the grace period with the cleanup/cutover of
the recovery db.

Protect changes to current_grace with the grace_mutex. We can still do
fetches against it for grace checks without the mutex however since
those are not ordered in any way with other activities. Give a
current_grace value of 0 a special meaning that the grace period
is done.

When we go to lift the grace period clean out the recovery db, and
set current_grace to 0.

Change-Id: Ia396978971d4a07ff42fd9dbd5da65c1af0450ba
Signed-off-by: Jeff Layton <jlayton@redhat.com>