Update patch set 2
Patch Set 2:

One question on this is what happens if the state_t related opens are above the high water mark but there are no global opens? Does that cause the LRU thread to run aggressively but not actually accomplish anything?

We should count the state_t and not allow more than a certain number (could be based of high water mark, though that should be able to be set low if most of the exports are expected to NOT use open fd such as FSAL_PROXY).

With 2.6 coming to a close, we need to balance what needs to be done to fix current performance issues for products that can't wait for 2.7, but we may have to save the elegant fix for 2.7.

Patch-set: 2