NFS v4.1 - use cachethis indication from client
Change the name of use_drc to use_slot_cached_result to clarify
what is being directed.

Remove cache_used (it is a duplicate of res_cached inside
struct COMPOUND4res_extended.

Rename cached_res in compound_data_t to cached_result for
simplicity (now cscope will find both cached_result in the slot and
in the compound data).

Note that the cached result is a small structure, the array of
results for each compound op is referenced by pointer. The function
nfs4_Compound_Free does a deep free only if the response is not
cached, and thus, any place the slot cache needs to be cleaned up
is simply accomplished by clearing res_cached and calling
nfs4_Compound_Free.

The logic changes in this patch are confined to:

Protocols/NFS/nfs4_op_sequence.c which checks cachethis before
setting up the cache, and cleaning up the slot cache if cachethis is
false since it will no longer be referenced (and could cause
confusion on the next request that is cached).

Change-Id: Icf5d0fb9349b8eb2f6bb26103959b55918616226
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>