Consolidate COMPOUND structure checks and fix DESTROY_SESSION check
Correct the DESTROY_SESSION check if it's in a COMPOUND with
SEQUENCE, if it's for the same session, it must be the final op.

Apparently you can destroy as many sessions as you want in a single
COMPOUND so long as if you are destroying THAT session you do it
last... Which all makes sense and is cool.

Change-Id: I4c9320a5f9a79d79e8369f287824873bf8bf6a31
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>