fsal_rgw: support directory object as an export
FSAL RGW currently only support "/" or "bucket" as an export, I think we can also make a directory under a bucket as an export too, like:

EXPORT
{
  Export_ID=1;
  Path = "bucket1/dir1";
  Pseudo = "/";
  Access_Type = RW;
  Protocols = 4;
  Squash = No_Root_Squash;
  Transports = TCP;
  ...
}

The reason that I made this is, that I found if there is a huge number of objects inside a bucket(like millions), the "ls" operation to mount point will extremely slow because Ceph Rados will parse all objects under this bucket, which is unnecessary. So if it is possible to make a directory under a bucket as an export, we could divide a bucket into several directory, which will surely reduce "ls" operation cost.

Change-Id: I278b55fcba368ecf67e8bde010232c5b46f6dd8f
Signed-off-by: taoCH <sebastien.chen92@gmail.com>