Update patch set 1
Patch Set 1:

> I think we should handle this in the protocol layer so it is
 > handled for all FSALs.

I initially thought about having the change as you suggested in protocol layer. But the object type as SYMLINK gets identified first in some cases in FSAL_GPFS and then GPFS alloc_handle() is called. And later on when mdcache_new_entry() finds new handle is not needed it calls FSAL_GPFS release() where crash happens. The call to mdcache_new_entry() happens before the fsal_open2() returns to the protocol layer. So if we make a change at protocol layer then crash would still happen. The coredump for the crash is at https://paste.fedoraproject.org/paste/HbDVe624tbsLHtnbr42uMg
Another option could be to have the change in mdcache_open2() for 'state' object to check if object type is SYMLINK and return error in that case. That would help to avoid the crash as well we will have the change for all FSALs. Please let me know your comments.

Patch-set: 1