Update patch set 1
Patch Set 1:

> > Investigated the code, if this is handled in the mdcache_open2()
 > > just after calling the FSAL specfic open2(), then for object type
 > > SYMLINK we would have to set the error and would need to call the
 > > FSAL release() to release any newly allocated object. But this
 > > would lead to the same crash in release() as the FSAL specific
 > > open2() had already set 'u.file.share' and 'u.symlink.link_content'
 > > where both the 'u.symlink' structure and 'u.file' structure which
 > > are part of the same union 'u'.
 > > It seems like we need to take help of FSAL to detect object type
 > as
 > > SYMLINK as soon as possible and take some action. Please let me
 > > know know your comments.
 > 
 > How about doing this in open4_ex()? I am not sure if I am covering
 > everything though. open with create is not going to create a
 > symlink. fsal_lookup() is currently not passing attrs pointer. If
 > it passes and it happens to be a symlink, can we bailout from
 > there?

open4_ex already checks for symlink if the lookup succeeded, so somehow there is a path where the lookup did not succeed. The fsal_open2 helper path also checks for regular file if it has an object. mdcache probably should also check if it resolves a name to an object handle.

It would be good to understand the exact path to failure here, maybe there's a bug in the above code. But if genuinely we are getting to the FSAL to try and open a file, then the FSAL does indeed need to protect itself, and in that case, we probably need to do something in all FSALs.

Frank

Patch-set: 1