Update patch set 1
Patch Set 1: Code-Review+1

> 
 > Following is the code path for the crash that we observed.
 > nfs4_op_open()
 > --> open4_ex()
 > --> fsal_lookup() (fails, looked_up_file_obj=false)
 > --> fsal_open2()
 > --> open2_by_name()
 > --> mdcache_open2()
 > --> mdc_open2_by_name() (fails)
 > --> gpfs_open2() (success, opens a file descriptor)
 > --> GPFSFSAL_create2() (success)
 > --> alloc_handle() (success, allocates object, sets
 > 'u.symlink.link_content=NULL')
 > --> open_by_handle() (success, opens a file, sets 'u.file.share' to
 > valid value.
 > With this 'u.symlink.link_content' gets indirectly set as
 > 'u.symlink' and 'u.file' structures are part of same union 'u')
 > --> mdcache_alloc_and_check_handle()
 > --> mdcache_new_entry()
 > --> release() (calls gsh_free() for u.symlink.link_content as its
 > not NULL)
 > --> gsh_free() (crashes because 'u.symlink.link_content' doesn't
 > point to valid address)

So, if I'm reading this right, this is an actual open/create where the passed in attributes make the newly opened file a symlink?  This case can be handled by open4_ex(), fsal_open2(), or mdcache_open2().  gpfs_open2() should probably fail anyway, for safety.

I guess the other possibility is a race where open/create attempts to create a normal file, and another thread creates the symlink first.  In this case, gpfs_open2() must be able to detect it and fail.

So, I think this patch is correct, and the other FSALs will need to handle this case as well, but we should probably add the check for open/create higher in the stack (my preference would be fsal_open2()).

Patch-set: 1
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1