Update patch set 1
Patch Set 1:

(1 comment)

> >
 > > Following is the code path for the crash that we observed.
 > > nfs4_op_open()
 > > --> open4_ex()
 > > --> fsal_lookup() (fails, looked_up_file_obj=false)
 > > --> fsal_open2()
 > > --> open2_by_name()
 > > --> mdcache_open2()
 > > --> mdc_open2_by_name() (fails)
 > > --> gpfs_open2() (success, opens a file descriptor)
 > > --> GPFSFSAL_create2() (success)
 > > --> alloc_handle() (success, allocates object, sets
 > > 'u.symlink.link_content=NULL')
 > > --> open_by_handle() (success, opens a file, sets 'u.file.share'
 > to
 > > valid value.
 > > With this 'u.symlink.link_content' gets indirectly set as
 > > 'u.symlink' and 'u.file' structures are part of same union 'u')
 > > --> mdcache_alloc_and_check_handle()
 > > --> mdcache_new_entry()
 > > --> release() (calls gsh_free() for u.symlink.link_content as its
 > > not NULL)
 > > --> gsh_free() (crashes because 'u.symlink.link_content' doesn't
 > > point to valid address)
 > 
 > So, if I'm reading this right, this is an actual open/create where
 > the passed in attributes make the newly opened file a symlink? 
 > This case can be handled by open4_ex(), fsal_open2(), or
 > mdcache_open2().  gpfs_open2() should probably fail anyway, for
 > safety.
 > 
 > I guess the other possibility is a race where open/create attempts
 > to create a normal file, and another thread creates the symlink
 > first.  In this case, gpfs_open2() must be able to detect it and
 > fail.
 > 
 > So, I think this patch is correct, and the other FSALs will need to
 > handle this case as well, but we should probably add the check for
 > open/create higher in the stack (my preference would be
 > fsal_open2()).

Ok, so what's happening is the lookup at the mdcache layer is failing (I'm getting lost following the exact logic there, and I actually have a question if we are opening ourselves to ESTALE conditions or if there's any path where we can do a subfsal lookup followed by a subfsal open2 by handle when we would be better off just directly subfsal open2 by name), so we call the subfsal open2 by name. The subfsal could use O_NOFOLLOW to stop this immediately.

At the mdcache layer, if we get an open2 by name, we should actually call the subfsal open2 by name if we get an ESTALE from any attempt to open by handle. If we had a cached object by that name, and now it's actually been removed (thus someone throws an ESTALE), we should actually see if there happens to be a NEW file by that name on disk, or if it's a create, definitely create one.

fsal_open2 is way too late to do it, or we are adding even another layer of doing a lookup before actual open.

Part of the idea of the open2 method was to make a more atomic open by the FSAL, and all this lookup logic above it goes against that idea. I even have work in place to use Linux's recent ability to create unnamed files to even better do exclusive create in FSAL_VFS.

Patch-set: 1